Remove the flat data source for NSCollectionView based on review feedback
